requirements.txt
streamlit==1.31.0
pandas==1.2.3
requests==2.22.0
dataclasses==0.6
python-dotenv==1.0.1
replicate==0.24.0
black==23.12.1